refactor(client): Go package name/import fix for issue #1687
Relocated go packge "zbc" from "github.com/zeebe-io/zeebe/clients/go"
to "github.com/zeebe-io/zeebe/clients/go/zbc" directory, this would break
existing code, requiring to update import paths.
