+++ clients/go/tests/integration/cancelInstance_test.go
@@ -3,7 +3,7 @@ package integration
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/zeebe-io/zeebe/clients/go"
+	"github.com/zeebe-io/zeebe/clients/go/zbc"
 )
 
 var _ = Describe("CreateInstance", func() {
